# Automatic memory layout adaption for increased GPU performance