--- a/src/RadiationDetectorDSP.jl
+++ b/src/RadiationDetectorDSP.jl
@@ -53,4 +53,15 @@ include("trapezoidal_filter.jl")
 include("sg_filter.jl")
 include("zac_filter.jl")
 
+
+@static if !isdefined(Base, :get_extension)
+    using Requires
+end
+
+function __init__()
+    @static if !isdefined(Base, :get_extension)
+        @require CUDA = "052768ef-5323-5732-b1bb-66c8b64840ba" include("../ext/CUDAExt.jl")
+    end
+end
+
 end # module