--- a/src/array_utils.jl
+++ b/src/array_utils.jl
@@ -45,6 +45,16 @@ _similar_memlayout(A::AbstractArray{<:T}, ::Type{U}, sz::Dims) where {T<:Real,U<
 _similar_memlayout(A::LinearAlgebra.Transpose{<:T}, ::Type{U}, sz::Dims) where {T<:Real,U<:Real} = transpose(similar(A, U, reverse(sz)))
 
 
+_lazy_transpose(A::AbstractMatrix{T}) where {T<:Number} = transpose(A)
+_nonlazy_transpose(A::AbstractMatrix{T}) where {T<:Number} = copy(_lazy_transpose(A))
+
+_row_major(A::AbstractMatrix{T}) where {T<:Number} = _lazy_transpose(_nonlazy_transpose(A))
+_row_major(A::LinearAlgebra.Transpose{T}) where {T<:Number} = A
+
+_col_major(A::AbstractMatrix{T}) where {T<:Number} = A
+_col_major(A::LinearAlgebra.Transpose{T}) where {T<:Number} = _nonlazy_transpose(_lazy_transpose(A))
+
+
 # ToDo: Replace _to_same_device_as workaround as soon as ArrayInferface
 # and Adapt have proper support for computing devices: