--- a/src/first_order_iir.jl
+++ b/src/first_order_iir.jl
@@ -102,6 +102,8 @@ end
 end
 
 
+adapt_memlayout(::FirstOrderIIRInstance, ::GPU, A::AbstractArray{<:Number}) = _row_major(A)
+
 function bc_rdfilt!(
     outputs::ArrayOfSimilarVectors{<:RealQuantity},
     fi::FirstOrderIIRInstance,