CPU threads prefer to access the samples of each waveform in contiguous memory (leaving SIMD considerations aside). GPU threads (at least each NVIDIA warps) prefer to access each sample across several waveforms in a contiguous fashion if filters are based on sequential loops along the waveform (otherwise NVIDIA warps and the like can't do coalesced memory reads). But operations like FFT perform better with samples in contiguous memory on GPUs.

This PR introduces automatic memory layout changes based on device and waveform filter type.

Performance test on Julia v1.9.0-beta3, CUDA.jl v4.0.1, on NVIDIA A100:

```julia
using RadiationDetectorDSP, ArraysOfArrays, Random, BenchmarkTools, CUDA, CUDAKernels
nsamples = 10^4
nwaveforms = 10^5
wfs = nestedview(rand!(CuArray{Float32}(undef, nsamples, nwaveforms)));
flt = TrapezoidalChargeFilter(500, 100)
@benchmark flt.($wfs)
```

Before:

```
BenchmarkTools.Trial: 15 samples with 1 evaluation.
 Range (min … max):  341.918 ms … 355.852 ms  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     346.078 ms               ┊ GC (median):    0.00%
 Time  (mean ± σ):   346.459 ms ±   3.576 ms  ┊ GC (mean ± σ):  0.05% ± 0.18%
```

After:

```
BenchmarkTools.Trial: 204 samples with 1 evaluation.
 Range (min … max):  22.335 ms … 31.743 ms  ┊ GC (min … max): 0.00% … 10.23%
 Time  (median):     22.522 ms              ┊ GC (median):    0.00%
 Time  (mean ± σ):   24.481 ms ±  2.745 ms  ┊ GC (mean ± σ):  3.83% ±  4.77%
```

Note: Uses fast matrix transpositions on NVIDIA only so far.